requirements.txt
streamlit
rdkit-pypi
pillow
numpy==1.26.4
